Clean up README
Previously, we had a list of instructions on how to compile Emacs from
scratch, which is something that I am no longer interested in doing.

Now, we will use it as a sort of repository of ideas to
implement. They will be investigated and added to a changelog (or
deleted) as you go on adding stuff.